refactor(content): 진행 상황 분리
- 구입 금액 입력 부 -> FirstSection
- 그 이후 컨텐츠 -> NextSection
