+++ src/view/web/main/NextSection.js
@@ -1,0 +1,24 @@
+import WinningLottoContent from './WinningLottoContent.js';
+
+export default function NextStepArticle() {
+  const section = document.createElement('section');
+  section.classList.add('lotto-section');
+
+  const nextSection = document.createElement('section');
+  nextSection.setAttribute('id', 'step2');
+  const buyTotalText = document.createElement('span');
+  buyTotalText.setAttribute('id', 'total-buy-text');
+  const lottoTicketsContainer = document.createElement('section');
+  const ul = document.createElement('ul');
+  lottoTicketsContainer.setAttribute('id', 'lotto-tickets-container');
+
+  lottoTicketsContainer.appendChild(ul);
+
+  nextSection.appendChild(buyTotalText);
+  nextSection.appendChild(lottoTicketsContainer);
+  nextSection.appendChild(WinningLottoContent());
+
+  section.appendChild(nextSection);
+
+  return section;
+}
